Clean up errno handling
We DO NOT need to reset errno if we check the return value of a
system call, and only use errno if the call failed.

We DO need to reset errno if we use errno without a guarantee
that a specific system function failed. We should not assume any
s2n_* function sets errno on failure.

I left the errno=0 at the beginning of s2n_negotiate as a
defense against future mistakes.